src/requirements.txt
Django==1.9
djangorestframework==3.6.0
whitenoise==3.3.1
django-crispy-forms==1.7.0
django-model-utils==3.0.0
Pillow==5.0.0
PyYAML==3.12
Markdown==2.6.11
django-db-logger==0.1.6




